breaking change: c++: split rust_vec.h into rust_vec.h and rust_splice.h
