+++ macroslib/src/lib.rs
@@ -306,8 +306,16 @@ impl Generator {
                     code: include_str!("cpp/rust_vec.h").into(),
                 });
                 foreign_lang_helpers.push(SourceCode {
+                    id_of_code: "rust_slice.h".into(),
+                    code: include_str!("cpp/rust_slice.h").into(),
+                });
+                foreign_lang_helpers.push(SourceCode {
                     id_of_code: "rust_result.h".into(),
                     code: include_str!("cpp/rust_result.h").into(),
+                });
+                foreign_lang_helpers.push(SourceCode {
+                    id_of_code: "rust_foreign_slice_iter.hpp".into(),
+                    code: include_str!("cpp/rust_foreign_slice_iter.hpp").into(),
                 });
             }
         }
